installer: add uid detection
This commit adds UID detection to installer-v2. If it detects a UID
which is not 0, it will exit the installer and send messages to the
standard error. It will also display instructions for the end user on
how to solve the issue. The function is currently commented out in the
main function to make testing easier, but it should be uncommented for
the final release of installer-v2.

Signed-off-by: spacefault <74156492+spacefault@users.noreply.github.com>