installer: add confirm screen
This commit adds the confirm screen to the installer. Firstly, it will
ask the user to type in the string 'InstallCavemanLinux'. Then, it will
check if the user input matches the one in the script. If it is not
approved, it will ask again, and if it is approved it will do the
opposite.

Signed-off-by: spacefault <74156492+spacefault@users.noreply.github.com>